pyproject.toml
[tool.poetry]
name = "datebear"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^1.26.4"
pandas = "^2.2.2"
plotnine = "^0.13.4"
transformers = "^4.40.1"
jupyter = "^1.0.0"
torch = "^2.3.0"
sentence-transformers = "^2.7.0"
scikit-learn = "^1.4.2"
haystack-ai = "^2.2.3"
openai = "^1.34.0"
streamlit = "^1.35.0"
fastapi = "^0.111.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
